Fixed a bug with newtype JSON marshalling
I had added field names to Model and Host to make defaults in
LLMClient.Common

Did not test much and this broke how the Model type is serialized to
JSON for the POST to the LLM server. Decided to just String for the
defaults values of these two things because it's the type expected by
the optparse-applicative machinery anyway, there's already a Text pack
going on.