Using Text.pack in less places
There are a number of places where the code was turning String into Text
to then format back into a String with Formatting. No need for this with
the s/string formatter.

Also added some usage info about why this project was made because it
shares overlap with basic command-line features already in ollama.